Add ServiceAccount which can access Control API apiserver metrics
Configure ServiceAccount, long-lived token, and ClusterRoleBinding to
`system:monitoring` to allow the service account access to the apiserver
`/metrics` API endpoint.

We pick a very short name for the service account and secret, so that
there's a reasonable chance that the name won't get mangled when the
secret is synced from a vcluster instance into the host cluster, if the
control-api is deployed in a vcluster.

Additionally, we add annotation `vcluster.loft.sh/force-sync=true` to
the token secret to ensure that this secret is available in the host
cluster to allow us to scrape metrics.